Add movement factory and Deck-To-Waste movement
Movements can be treated as first-order functions, with the signature:

 f(Pile, Pile, Int): (Pile, Pile)

That is, functions applied to two piles and a number of cards, that
return a new tuple of piles. The MovementFactory is used to create such
functions with the restrictions of each movement.

Note that the applicacion of movement functions may throw exceptions if
some rule is not followed (e.g. picking from an empty Pile).